Introduce flags PF, SF, ZF
I didn't understand how CMP worked. I thought the "machine status word" was
a register (which I had lettered as C) which stored the result of the arithmetic
operation (in the case of CMP, subtraction).

Rather, it seems the result of CMP is discarded, and the "status word" is
actually a few distinct bits.